backend/requirements.txt
fastapi[all]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
gunicorn
python-multipart
pydantic-settings